refactor: Remove ModelWrapper class.
