+++ trip-review-summary/src/review_summary/query/fetch_data/fetch_relationship.py
